Provisioning: refactor integration helper test out of the test package (#119410)
* chore: move provisioning test helper to common package

Move helper_test.go to common/testing.go to make it importable outside
the test package.

Made-with: Cursor

* chore: refactor integration tests to use common package helper

Update all provisioning integration test files to import the shared
helper from the new common package instead of using the inline helper.

Made-with: Cursor

* chore: update integration test to use new helper

* chore: update latest integration tests to use the separate package

* chore: revert back missing change after conflict merge
