+++ pkg/tests/apis/provisioning/classic_delete_test.go
@@ -11,6 +11,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/util/testutil"
 )
 
@@ -22,13 +23,13 @@ import (
 func TestIntegrationProvisioning_ClassicDashboardDeletion(t *testing.T) {
 	testutil.SkipIntegrationTestInShortMode(t)
 
-	helper := runGrafana(t)
+	helper := common.RunGrafana(t)
 
 	t.Run("classic dashboard deleted via filesystem removal and sync", func(t *testing.T) {
 		const repo = "classic-delete-repo"
 		repoPath := filepath.Join(helper.ProvisioningPath, repo)
 
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:   repo,
 			Path:   repoPath,
 			Target: "folder",
@@ -72,7 +73,7 @@ func TestIntegrationProvisioning_ClassicDashboardDeletion(t *testing.T) {
 		err = os.WriteFile(filepath.Join(repoPath, "inline-classic.json"), classicDashboard, 0o600)
 		require.NoError(t, err)
 
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:                   repo,
 			Path:                   repoPath,
 			Target:                 "folder",
@@ -112,7 +113,7 @@ func TestIntegrationProvisioning_ClassicDashboardDeletion(t *testing.T) {
 		err = os.WriteFile(filepath.Join(repoPath, "status-check.json"), classicDashboard, 0o600)
 		require.NoError(t, err)
 
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:                   repo,
 			Path:                   repoPath,
 			Target:                 "folder",
@@ -178,7 +179,7 @@ func TestIntegrationProvisioning_ClassicDashboardDeletion(t *testing.T) {
 		err = os.WriteFile(filepath.Join(repoPath, "k8s.json"), k8sDashboard, 0o600)
 		require.NoError(t, err)
 
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:                   repo,
 			Path:                   repoPath,
 			Target:                 "folder",
@@ -199,7 +200,7 @@ func TestIntegrationProvisioning_ClassicDashboardDeletion(t *testing.T) {
 				}
 			}
 			assert.Equal(collect, 2, count, "both dashboards should exist")
-		}, waitTimeoutDefault, waitIntervalDefault, "both dashboards should be created")
+		}, common.WaitTimeoutDefault, common.WaitIntervalDefault, "both dashboards should be created")
 
 		// Delete both files
 		err = os.Remove(filepath.Join(repoPath, "classic.json"))
