+++ pkg/tests/apis/provisioning/fieldselector_test.go
@@ -9,6 +9,7 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/util/testutil"
 )
 
@@ -18,7 +19,7 @@ import (
 func TestIntegrationProvisioning_RepositoryFieldSelector(t *testing.T) {
 	testutil.SkipIntegrationTestInShortMode(t)
 
-	helper := runGrafana(t)
+	helper := common.RunGrafana(t)
 	ctx := context.Background()
 
 	// Create multiple repositories for testing
@@ -106,7 +107,7 @@ func TestIntegrationProvisioning_RepositoryFieldSelector(t *testing.T) {
 func TestIntegrationProvisioning_JobFieldSelector(t *testing.T) {
 	testutil.SkipIntegrationTestInShortMode(t)
 
-	helper := runGrafana(t)
+	helper := common.RunGrafana(t)
 	ctx := context.Background()
 
 	// Create a repository to trigger jobs
@@ -130,7 +131,7 @@ func TestIntegrationProvisioning_JobFieldSelector(t *testing.T) {
 	}
 
 	// Trigger first job
-	body1 := asJSON(job1Spec)
+	body1 := common.AsJSON(job1Spec)
 	result1 := helper.AdminREST.Post().
 		Namespace("default").
 		Resource("repositories").
