+++ pkg/tests/apis/provisioning/files_quota_test.go
@@ -11,6 +11,7 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/tests/testinfra"
 	"github.com/grafana/grafana/pkg/util/testutil"
 )
@@ -20,12 +21,12 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 
 	t.Run("no quota configured allows create and update via files endpoint", func(t *testing.T) {
 		// No quota limit = unlimited, both POST (create) and PUT (update) should succeed
-		helper := runGrafana(t)
+		helper := common.RunGrafana(t)
 		ctx := context.Background()
 
 		const repo = "files-quota-unlimited-repo"
 		repoPath := filepath.Join(helper.ProvisioningPath, repo)
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:   repo,
 			Path:   repoPath,
 			Target: "folder",
@@ -67,10 +68,10 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 		require.Equal(t, http.StatusOK, updateStatusCode, "should return 200 OK for update")
 
 		// POST with originalPath (move) should succeed
-		resp := helper.postFilesRequest(t, repo, filesPostOptions{
-			targetPath:   "moved-dashboard.json",
-			originalPath: "new-dashboard.json",
-			message:      "move dashboard",
+		resp := helper.PostFilesRequest(t, repo, common.FilesPostOptions{
+			TargetPath:   "moved-dashboard.json",
+			OriginalPath: "new-dashboard.json",
+			Message:      "move dashboard",
 		})
 		defer resp.Body.Close() //nolint:errcheck
 		require.Equal(t, http.StatusOK, resp.StatusCode, "POST (move) should succeed when no quota is configured")
@@ -78,14 +79,14 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 
 	t.Run("within quota allows create and update via files endpoint", func(t *testing.T) {
 		// With folder target: 1 dashboard + 1 folder = 2 resources, limit 10 → within quota
-		helper := runGrafana(t, func(opts *testinfra.GrafanaOpts) {
+		helper := common.RunGrafana(t, func(opts *testinfra.GrafanaOpts) {
 			opts.ProvisioningMaxResourcesPerRepository = 10
 		})
 		ctx := context.Background()
 
 		const repo = "files-quota-within-repo"
 		repoPath := filepath.Join(helper.ProvisioningPath, repo)
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:   repo,
 			Path:   repoPath,
 			Target: "folder",
@@ -126,10 +127,10 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 		require.Equal(t, http.StatusOK, updateStatusCode, "should return 200 OK for update")
 
 		// POST with originalPath (move) should succeed
-		resp := helper.postFilesRequest(t, repo, filesPostOptions{
-			targetPath:   "moved-dashboard.json",
-			originalPath: "new-dashboard.json",
-			message:      "move dashboard",
+		resp := helper.PostFilesRequest(t, repo, common.FilesPostOptions{
+			TargetPath:   "moved-dashboard.json",
+			OriginalPath: "new-dashboard.json",
+			Message:      "move dashboard",
 		})
 		defer resp.Body.Close() //nolint:errcheck
 		require.Equal(t, http.StatusOK, resp.StatusCode, "POST (move) should succeed when within quota")
@@ -137,14 +138,14 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 
 	t.Run("quota reached blocks create but allows update via files endpoint", func(t *testing.T) {
 		// With folder target: 1 dashboard + 1 folder = 2 resources, limit 2 → exactly at limit (reached)
-		helper := runGrafana(t, func(opts *testinfra.GrafanaOpts) {
+		helper := common.RunGrafana(t, func(opts *testinfra.GrafanaOpts) {
 			opts.ProvisioningMaxResourcesPerRepository = 2
 		})
 		ctx := context.Background()
 
 		const repo = "files-quota-reached-repo"
 		repoPath := filepath.Join(helper.ProvisioningPath, repo)
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:   repo,
 			Path:   repoPath,
 			Target: "folder",
@@ -184,10 +185,10 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 		require.Equal(t, http.StatusOK, updateStatusCode, "should return 200 OK for update")
 
 		// POST with originalPath (move) should succeed
-		resp := helper.postFilesRequest(t, repo, filesPostOptions{
-			targetPath:   "moved-dashboard.json",
-			originalPath: "dashboard1.json",
-			message:      "move dashboard",
+		resp := helper.PostFilesRequest(t, repo, common.FilesPostOptions{
+			TargetPath:   "moved-dashboard.json",
+			OriginalPath: "dashboard1.json",
+			Message:      "move dashboard",
 		})
 		defer resp.Body.Close() //nolint:errcheck
 		require.Equal(t, http.StatusOK, resp.StatusCode, "POST (move) should succeed when quota is reached")
@@ -195,12 +196,12 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 
 	t.Run("quota exceeded blocks both create and update via files endpoint", func(t *testing.T) {
 		// With folder target: 2 dashboards + 1 folder = 3 resources, limit 1 → exceeded
-		helper := runGrafana(t)
+		helper := common.RunGrafana(t)
 		ctx := context.Background()
 
 		const repo = "files-quota-exceeded-repo"
 		repoPath := filepath.Join(helper.ProvisioningPath, repo)
-		helper.CreateRepo(t, TestRepo{
+		helper.CreateRepo(t, common.TestRepo{
 			Name:   repo,
 			Path:   repoPath,
 			Target: "folder",
@@ -246,14 +247,14 @@ func TestIntegrationProvisioning_FilesQuotaEnforcement(t *testing.T) {
 		require.True(t, apierrors.IsForbidden(result.Error()), "error should be Forbidden, got: %v", result.Error())
 
 		// POST with originalPath (move) should also be blocked
-		resp := helper.postFilesRequest(t, repo, filesPostOptions{
-			targetPath:   "moved-dashboard.json",
-			originalPath: "dashboard1.json",
-			message:      "move dashboard",
+		resp := helper.PostFilesRequest(t, repo, common.FilesPostOptions{
+			TargetPath:   "moved-dashboard.json",
+			OriginalPath: "dashboard1.json",
+			Message:      "move dashboard",
 		})
 		defer resp.Body.Close() //nolint:errcheck
 		body, err := io.ReadAll(resp.Body)
 		require.NoError(t, err)
-		require.Equal(t, http.StatusForbidden, resp.StatusCode, "POST (move) should be blocked when quota is exceeded, body: %s", string(body))
+		require.Equal(t, http.StatusForbidden, resp.StatusCode, "POST (move) should be blocked when quota is exceeded, Body:         %s", string(body))
 	})
 }
