+++ pkg/tests/apis/provisioning/job_conflict_test.go
@@ -10,6 +10,7 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/tests/testinfra"
 	"github.com/grafana/grafana/pkg/util/testutil"
 )
@@ -20,7 +21,7 @@ func TestIntegrationProvisioning_JobConflict(t *testing.T) {
 	testutil.SkipIntegrationTestInShortMode(t)
 
 	// disable the controllers so the jobs don't get auto-processed
-	helper := runGrafana(t, func(opts *testinfra.GrafanaOpts) {
+	helper := common.RunGrafana(t, func(opts *testinfra.GrafanaOpts) {
 		opts.DisableControllers = true
 	})
 	ctx := context.Background()
