+++ pkg/tests/apis/provisioning/jobs_auth_test.go
@@ -9,17 +9,18 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/util/testutil"
 )
 
 func TestIntegrationProvisioning_JobsAuthorization(t *testing.T) {
 	testutil.SkipIntegrationTestInShortMode(t)
 
-	helper := runGrafana(t)
+	helper := common.RunGrafana(t)
 	ctx := context.Background()
 
 	const repo = "jobs-auth-test"
-	testRepo := TestRepo{
+	testRepo := common.TestRepo{
 		Name:               repo,
 		Target:             "folder",
 		Copies:             map[string]string{},
@@ -63,7 +64,7 @@ func TestIntegrationProvisioning_JobsAuthorization(t *testing.T) {
 	})
 
 	t.Run("admin can create job", func(t *testing.T) {
-		body := asJSON(provisioning.JobSpec{
+		body := common.AsJSON(provisioning.JobSpec{
 			Action: provisioning.JobActionPull,
 			Pull:   &provisioning.SyncJobOptions{},
 		})
@@ -85,7 +86,7 @@ func TestIntegrationProvisioning_JobsAuthorization(t *testing.T) {
 	})
 
 	t.Run("editor can create job", func(t *testing.T) {
-		body := asJSON(provisioning.JobSpec{
+		body := common.AsJSON(provisioning.JobSpec{
 			Action: provisioning.JobActionPull,
 			Pull:   &provisioning.SyncJobOptions{},
 		})
@@ -107,7 +108,7 @@ func TestIntegrationProvisioning_JobsAuthorization(t *testing.T) {
 	})
 
 	t.Run("viewer cannot create job", func(t *testing.T) {
-		body := asJSON(provisioning.JobSpec{
+		body := common.AsJSON(provisioning.JobSpec{
 			Action: provisioning.JobActionPull,
 			Pull:   &provisioning.SyncJobOptions{},
 		})
