+++ pkg/tests/apis/provisioning/librarypanels_test.go
@@ -9,6 +9,7 @@ import (
 	foldersV1 "github.com/grafana/grafana/apps/folder/pkg/apis/folder/v1beta1"
 	"github.com/grafana/grafana/pkg/apimachinery/utils"
 	"github.com/grafana/grafana/pkg/services/accesscontrol/resourcepermissions"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/util/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -22,8 +23,8 @@ import (
 func TestIntegrationLibraryPanels_ProvisionedFolders(t *testing.T) {
 	testutil.SkipIntegrationTestInShortMode(t)
 
-	helper := runGrafana(t)
-	helper.CreateRepo(t, TestRepo{
+	helper := common.RunGrafana(t)
+	helper.CreateRepo(t, common.TestRepo{
 		Name:            "test-repo",
 		Target:          "folder",
 		ExpectedFolders: 1,
@@ -45,7 +46,7 @@ func TestIntegrationLibraryPanels_ProvisionedFolders(t *testing.T) {
 			},
 		}
 		libraryElementURL := "/api/library-elements"
-		libraryElementData, code, err := postHelper(t, *helper.K8sTestHelper, libraryElementURL, libraryElement, helper.Org1.Admin)
+		libraryElementData, code, err := common.PostHelper(t, *helper.K8sTestHelper, libraryElementURL, libraryElement, helper.Org1.Admin)
 		require.Error(t, err)
 		require.Equal(t, http.StatusConflict, code)
 		require.NotNil(t, libraryElementData)
@@ -85,7 +86,7 @@ func TestIntegrationLibraryPanels_ProvisionedFolders(t *testing.T) {
 			},
 		}
 		libraryElementURL := "/api/library-elements"
-		libraryElementData, code, err := postHelper(t, *helper.K8sTestHelper, libraryElementURL, libraryElement, helper.Org1.Admin)
+		libraryElementData, code, err := common.PostHelper(t, *helper.K8sTestHelper, libraryElementURL, libraryElement, helper.Org1.Admin)
 		require.NoError(t, err)
 		require.Equal(t, http.StatusOK, code)
 		require.NotNil(t, libraryElementData)
@@ -107,7 +108,7 @@ func TestIntegrationLibraryPanels_ProvisionedFolders(t *testing.T) {
 			"folderUid": managedFolderName,
 		}
 		patchLibraryElementURL := fmt.Sprintf("/api/library-elements/%f", +res["id"].(float64))
-		newLibraryElement, code, err := patchHelper(t, *helper.K8sTestHelper, patchLibraryElementURL, updatedLibraryElement, helper.Org1.Admin)
+		newLibraryElement, code, err := common.PatchHelper(t, *helper.K8sTestHelper, patchLibraryElementURL, updatedLibraryElement, helper.Org1.Admin)
 		require.Error(t, err)
 		require.Equal(t, http.StatusConflict, code)
 		require.NotNil(t, newLibraryElement)
@@ -117,8 +118,8 @@ func TestIntegrationLibraryPanels_ProvisionedFolders(t *testing.T) {
 
 func TestIntegrationLibraryPanels_UnprovisionedFolders(t *testing.T) {
 	const repo = "test-repo"
-	helper := runGrafana(t)
-	helper.CreateRepo(t, TestRepo{
+	helper := common.RunGrafana(t)
+	helper.CreateRepo(t, common.TestRepo{
 		Name:            repo,
 		Target:          "folder",
 		ExpectedFolders: 1,
@@ -167,7 +168,7 @@ func TestIntegrationLibraryPanels_UnprovisionedFolders(t *testing.T) {
 			},
 		}
 		libraryElementURL := "/api/library-elements"
-		libraryElementData, code, err := postHelper(t, *helper.K8sTestHelper, libraryElementURL, libraryElement, helper.Org1.Admin)
+		libraryElementData, code, err := common.PostHelper(t, *helper.K8sTestHelper, libraryElementURL, libraryElement, helper.Org1.Admin)
 		require.NoError(t, err)
 		require.Equal(t, http.StatusOK, code)
 		require.NotNil(t, libraryElementData)
